+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
+## Unreleased
+
+### Added
+
+* add `tools firehose-prometheus-exporter` to use as a client-side monitoring tool of a firehose endpoint
+
 ## v1.2.2
 
 * Hotfix 'nil pointer' panic when saving uninitialized cache